drfcommerce/drfcommerce/tests/conftest.py
from pytest_factoryboy import register
from rest_framework.test import APIClient
from .factories import (
    CategoryFactory,
    BrandFactory,
    ProductFactory,
    ProductLineFactory,
)
import pytest

register(CategoryFactory)
register(BrandFactory)
register(ProductFactory)
register(ProductLineFactory)


@pytest.fixture
def api_client():
    return APIClient
