drfcommerce/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
click==8.1.6
colorama==0.4.6
coverage==7.2.7
Django==3.2.20
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==18.13.0
flake8==5.0.4
importlib-metadata==4.2.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
pluggy==1.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.15.1
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typed-ast==1.5.5
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0
